% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_xml_nodeset}
\alias{bind_xml_nodeset}
\title{Convert an xml_nodeset containing <a> elements into a tbl_df}
\usage{
bind_xml_nodeset(xmln)
}
\arguments{
\item{xmln}{xml_nodeset produced by e.g. \code{html_nodes(page, 'a')}}
}
\value{
A tbl_df of information derived from the XML attributes of a nodeset.
}
\description{
Convert an xml_nodeset containing <a> elements into a tbl_df
}
