% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_json}
\alias{convert_to_json}
\title{Convert a named list to JSON}
\usage{
convert_to_json(a_list, pretty = TRUE, auto_unbox = TRUE, ...)
}
\arguments{
\item{a_list}{An R object to convert to JSON}

\item{pretty}{Add whitespace to JSON output? Default: TRUE [argument to \code{jsonlite::toJSON}]}

\item{auto_unbox}{Automatically "unbox" vectors of length one? Default: TRUE [argument to \code{jsonlite::toJSON}]}

\item{...}{Additional arguments to \code{jsonlite::toJSON}}
}
\value{
A character string formatted as JSON
}
\description{
The main reason for including this wrapper method is to make roxygen aware of jsonlite dependency
}
