% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GDS.R
\name{create_GDS}
\alias{create_GDS}
\title{Create GDS Dataset}
\usage{
create_GDS(...)
}
\arguments{
\item{...}{Arguments to \code{parse_GDS}}
}
\value{
TRUE if successful
}
\description{
Create GDS Dataset
}
