% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_KST13.R
\name{create_KST13}
\alias{create_KST13}
\title{Create Keys to Soil Taxonomy (13th edition) Datasets}
\usage{
create_KST13(...)
}
\arguments{
\item{...}{arguments passed to `download_KST`}
}
\value{
TRUE if successful
}
\description{
Create Keys to Soil Taxonomy (13th edition) Datasets
}
