% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NSSH.R
\name{create_NSSH}
\alias{create_NSSH}
\title{Create NSSH Dataset}
\usage{
create_NSSH(...)
}
\arguments{
\item{...}{Arguments to \code{parse_nssh_index}}
}
\value{
TRUE if successful
}
\description{
Create NSSH Dataset
}
