% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nssh_index.R
\name{create_nssh_index}
\alias{create_nssh_index}
\title{Parse the National Soil Survey Handbook (NSSH) Table of Contents to get eDirectives links}
\usage{
create_nssh_index(
  url = NULL,
  ignore.headers = "Amendments To Soil Taxonomy",
  outpath = "./inst/extdata/NSSH",
  download_pdf = "ifneeded",
  output_types = c("txt", "html"),
  keep_pdf = FALSE
)
}
\arguments{
\item{url}{A URL to parse for Table of Contents information.}

\item{ignore.headers}{A character vector of h3 level headers to ignore on the NSSH Table of contents webpage.}

\item{outpath}{A directory path to create "inst/extdata/NSSH" folder structure.}

\item{download_pdf}{Download official PDF files from eDirectives? default: "ifneeded"; options: TRUE/FALSE}

\item{output_types}{Options include \code{c("txt","html")} for processed PDF files.}

\item{keep_pdf}{Keep PDF files after processing TXT?}
}
\value{
A data.frame object containing link, part and section information for the NSSH. A directory "inst/extdata/NSSH" is created in \code{outpath} (Default: "./inst/extdata/NSSH/") with a numeric subfolder for each part in the NSSH.
}
\description{
This provides a basic framework and folder structure for assets that are part of the NCSS standards.
}
\details{
Hardcoded with \code{ignore.headers = "Part 615 – Amendments To Soil Taxonomy"}; TODO: set this to NULL when webpage is updated. Default URL: https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/ref/?cid=nrcs142p2_054240
}
