% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_OSD.R
\name{create_OSD}
\alias{create_OSD}
\title{Create OSD Dataset}
\usage{
create_OSD(..., download = TRUE)
}
\arguments{
\item{...}{not used}

\item{download}{_logical_. Download OSD Snapshot from OSDRegistry? Default: `TRUE`}
}
\value{
TRUE if successful
}
\description{
Create OSD Dataset
}
