% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PLANTS.R
\name{create_PLANTS}
\alias{create_PLANTS}
\title{Create PLANTS Dataset}
\usage{
create_PLANTS(...)
}
\arguments{
\item{...}{Arguments to `create_*()` function. Only `outpath` supported at this time.}
}
\value{
TRUE if successful
}
\description{
Create PLANTS Dataset
}
