% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset.R
\name{.read_asset}
\alias{.read_asset}
\title{Internal method for reading assets from known filetypes}
\usage{
.read_asset(fname)
}
\arguments{
\item{fname}{A (singular) file name for an asset}
}
\value{
A data.frame or similar data object
}
\description{
Internal method for reading assets from known filetypes
}
