% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset.R
\name{get_assets}
\alias{get_assets}
\title{Fetch the assets associated with a specified dataset folder and name}
\usage{
get_assets(folder, name)
}
\arguments{
\item{folder}{A folder name (e.g. \code{"NSSH"})}

\item{name}{A regular expression matching target asset file names (e.g. \code{"index"})}
}
\value{
A list, 1 list element per data source / file matching folder and name conditions
}
\description{
Fetch the assets associated with a specified dataset folder and name
}
