% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{get_create_functions}
\alias{get_create_functions}
\title{Get \code{create_} functions used to build SoilKnowledgeBase}
\usage{
get_create_functions()
}
\value{
A vector of functions with \code{create_} prefix in name
}
\description{
Get \code{create_} functions used to build SoilKnowledgeBase
}
