% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logmsg}
\alias{logmsg}
\title{Write a character vector of format strings to a log file}
\usage{
logmsg(logfile, fmt, ..., sep = "\\n")
}
\arguments{
\item{logfile}{path to log file}

\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{values to be passed into fmt. As in \code{sprintf}}

\item{sep}{argument to \code{cat}; default: \code{"\n"}}
}
\value{
(invisible) formatted messages
}
\description{
Write a character vector of format strings to a log file
}
