% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_OSD.R
\name{osd_to_json}
\alias{osd_to_json}
\title{Convert OSD plaintext to JSON using NSSH structural elements}
\usage{
osd_to_json(
  logfile = file.path(output_dir, "OSD.log"),
  input_dir = "OSD",
  pattern = "txt",
  output_dir = "inst/extdata/OSD",
  osd_files = NULL
)
}
\arguments{
\item{logfile}{Path to log file; default: \code{file.path(output_dir, "OSD/OSD.log")}}

\item{input_dir}{Default: \code{'OSD'}; files matching pattern are listed recursively}

\item{pattern}{Argument passed to \code{list.files} when \code{osd_files} is not specified}

\item{output_dir}{Default: \code{'inst/extdata/OSD'}; folder to create alphabetical folder structure with JSON files}

\item{osd_files}{Default \code{NULL}; Optional over-ride vector of file names for testing}
}
\value{
A logical vector equal in length to the number of input files.
}
\description{
Convert OSD plaintext to JSON using NSSH structural elements
}
