% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GDS.R
\name{parse_GDS}
\alias{parse_GDS}
\title{parse_GDS}
\usage{
parse_GDS(
  outpath = "./inst/extdata",
  download_pdf = "ifneeded",
  output_types = c("txt"),
  keep_pdf = FALSE
)
}
\arguments{
\item{outpath}{A directory path to create "inst/extdata/NSSH" folder structure.}

\item{download_pdf}{Download official PDF file? default: "ifneeded"; options: TRUE/FALSE}

\item{output_types}{Options include \code{c("txt","html")} for processed PDF files.}

\item{keep_pdf}{Keep PDF files after processing TXT?}
}
\description{
parse_GDS
}
