% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NSSH.R
\name{parse_NSSH}
\alias{parse_NSSH}
\title{Parse a Part/Subpart TXT file from the National Soil Survey Handbook}
\usage{
parse_NSSH(a_part, a_subpart)
}
\arguments{
\item{a_part}{Part number (a three digit integer, starting with 6)}

\item{a_subpart}{Subpart letter (A or B)}
}
\value{
TRUE if succesful
}
\description{
Parse a Part/Subpart TXT file from the National Soil Survey Handbook
}
