% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NSSH.R
\name{parse_nssh_index}
\alias{parse_nssh_index}
\title{Parse the National Soil Survey Handbook (NSSH) Table of Contents to get eDirectives links}
\usage{
parse_nssh_index(
  logfile = file.path(outpath, "NSSH/NSSH.log"),
  nssh_url = NULL,
  ignore.headers = NULL,
  outpath = "./inst/extdata",
  download_pdf = "ifneeded",
  output_types = c("txt", "html"),
  keep_pdf = FALSE,
  ...
)
}
\arguments{
\item{logfile}{Path to log file; default \code{file.path(outpath, "NSSH/NSSH.log")}}

\item{nssh_url}{A URL to parse for Table of Contents information.}

\item{ignore.headers}{A character vector of h3 level headers to ignore on the NSSH Table of contents webpage.}

\item{outpath}{A directory path to create "inst/extdata/NSSH" folder structure.}

\item{download_pdf}{Download official PDF files from eDirectives? default: "ifneeded"; options: TRUE/FALSE}

\item{output_types}{Options include \code{c("txt","html")} for processed PDF files.}

\item{keep_pdf}{Keep PDF files after processing TXT?}

\item{...}{Additional arguments to `curl::curl_download()`}
}
\value{
A data.frame object containing link, part and section information for the NSSH. A directory "inst/extdata/NSSH" is created in \code{outpath} (Default: "./inst/extdata/NSSH/") with a numeric subfolder for each part in the NSSH.
}
\description{
\code{parse_nssh_index} provides a basic framework and folder structure for assets that are part of the National Soil Survey Handbook (NSSH) a key part of National Cooperative Soil Survey (NCSS) standards.
}
