% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NSSH.R
\name{parse_nssh_part}
\alias{parse_nssh_part}
\title{Parse headers and line positions by NSSH Part and Subpart}
\usage{
parse_nssh_part(number, subpart, outpath = "./inst/extdata")
}
\arguments{
\item{number}{Vector of part number(s) e.g. \code{600:614}}

\item{subpart}{Vector of subpart characters e.g. \code{"A"}}

\item{outpath}{A directory path to create "inst/extdata/NSSH" folder structure in}
}
\value{
A data.frame containing line numbers corresponding to NSSH part and subpart headers.
}
\description{
Parse headers and line positions by NSSH Part and Subpart
}
