% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nssh_index.R
\name{refresh}
\alias{refresh}
\title{Refresh the entire SoilKnowledgeBase inst/extdata/ folder}
\usage{
refresh(...)
}
\arguments{
\item{...}{Arguments passed to \code{create_nssh_index}}
}
\value{
Raw data and parsed files written to inst/extdata/ subfolders.
}
\description{
Refresh the entire SoilKnowledgeBase inst/extdata/ folder
}
