% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_OSD.R
\name{validateOSD}
\alias{validateOSD}
\title{Validate OSD using NSSH standards}
\usage{
validateOSD(logfile, filepath)
}
\arguments{
\item{logfile}{Path to log file}

\item{filepath}{Path to a single plain text file containing OSD narrative}
}
\value{
A nested list containing OSD structure for specified file path
}
\description{
Validate OSD using NSSH standards
}
