% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_KST.R
\name{create_KST}
\alias{create_KST}
\title{Create Keys to Soil Taxonomy Datasets}
\usage{
create_KST(...)
}
\arguments{
\item{...}{arguments passed to `download_KST`}
}
\value{
TRUE if successful
}
\description{
Create Keys to Soil Taxonomy Datasets
}
