% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nssh_part.R
\name{parse_nssh_part}
\alias{parse_nssh_part}
\title{Parse headers and line positions by NSSH Part and Subpart}
\usage{
parse_nssh_part(number, subpart)
}
\arguments{
\item{number}{Vector of part number(s) e.g. \code{600:614}}

\item{subpart}{Vector of subpart characters e.g. \code{"A"}}
}
\value{
A data.frame containing line numbers corresponding to NSSH part and subpart headers.
}
\description{
Parse headers and line positions by NSSH Part and Subpart
}
