% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh the entire SoilKnowledgeBase inst/extdata/ folder}
\usage{
refresh(...)
}
\arguments{
\item{...}{Arguments passed to \code{parse_nssh_index} or similar index-creating methods}
}
\value{
TRUE if all \code{create_} functions returned TRUE.
}
\description{
This method calls all functions defined in the SoilKnowledgeBase package with the prefix \code{create_} in the function name.
}
